+++ tests/python/pants_test/tasks/BUILD
@@ -404,16 +404,30 @@ python_tests(
 )
 
 python_tests(
-  name = 'jar_publish',
-  sources = ['test_jar_publish.py'],
+  name = 'jar_artifact_publish',
+  sources = ['test_jar_artifact_publish.py'],
   dependencies = [
     '3rdparty/python:mock',
     ':base',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/tasks:jar_publish',
+    'src/python/pants/backend/jvm/tasks:jar_artifact_publish',
     'src/python/pants/base:build_file_aliases',
     'src/python/pants/base:source_root',
+    'src/python/pants/scm:scm',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:dirutil',
+    ],
+)
+
+python_tests(
+  name = 'jar_publish',
+  sources = ['test_jar_publish.py'],
+  dependencies = [
+    '3rdparty/python:mock',
+    ':jar_artifact_publish',
+    'src/python/pants/backend/jvm/tasks:jar_publish',
+    'src/python/pants/base:exceptions',
     'src/python/pants/scm:scm',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
